graph-tests: move `docker` and other test utilities to lib
