+++ tests/Cargo.toml
@@ -5,7 +5,6 @@ edition.workspace = true
 
 [dependencies]
 port_check = "0.1.5"
-anyhow = "1.0"
 futures = { version = "0.3", features = ["compat"] }
 graph = { path = "../graph" }
 tokio = { version = "1.16.1", features = ["rt", "macros", "process"] }
@@ -19,12 +18,13 @@ graph-store-postgres = { path = "../store/postgres" }
 slog = { version = "2.7.0", features = ["release_max_level_trace", "max_level_trace"] }
 graphql-parser = "0.4.0"
 hex = "0.4.3"
+bollard = "0.10"
+tokio-stream = "0.1"
+anyhow = "1.0.66"
 
 [dev-dependencies]
-bollard = "0.10"
-anyhow = "1.0.66"
 lazy_static = "1.4.0"
-tokio-stream = "0.1"
 serde_yaml = "0.8"
 cid = "0.8.6"
 graph-chain-near = { path = "../chain/near" }
+
