+++ tests/src/lib.rs
@@ -1,2 +1,5 @@
+pub mod docker;
 pub mod fixture;
 pub mod helpers;
+pub mod node;
+pub mod setup;
